Code refactor and welcome page title update
