+++ media/welcomePage/style.css
@@ -1,0 +1,249 @@
+.playbookGenerationContainer {
+	box-sizing: border-box;
+	line-height: 22px;
+	position: relative;
+	overflow: hidden;
+	height: inherit;
+	width: 100%;
+	user-select: initial;
+	-webkit-user-select: initial;
+	outline: none;
+}
+
+a .codicon {
+	font-size: 20px;
+}
+
+.playbookGenerationContainer.loading {
+	display: none;
+}
+
+.playbookGenerationContainer a {
+	text-decoration: none;
+}
+
+.playbookGenerationContainer img {
+	max-width: 100%;
+	max-height: 100%;
+	object-fit: contain;
+	pointer-events: none;
+}
+
+#system-readiness {
+	display: flex;
+	flex-direction: row;
+	margin-top: 14px;
+	background-color: var(--vscode-input-background);
+}
+
+.codicon.codicon-pass {
+	margin: 10px;
+	color: green;
+}
+
+
+#system-readiness-error {
+	display: flex;
+	flex-direction: row;
+	margin-top: 14px;
+	background-color: var(--vscode-input-background);
+}
+
+.codicon.codicon-error {
+	margin: 10px;
+	color: red;
+}
+
+.system-description {
+	padding: 0px;
+	margin: 6px;
+}
+
+#system-check {
+	display: flex;
+	align-items: start;
+	flex-direction: column;
+}
+
+#install-status {
+	padding: 0px 10px 0px 10px;
+	width: fit-content;
+	background-color: var(--vscode-input-background);
+}
+
+.found::before {
+	color: darkgreen;
+	content: "✓ ";
+}
+
+.not-found::before {
+	color: red;
+	content: "✗ ";
+}
+
+.not-found-optional::before {
+	color: gold;
+	content: "⚠ ";
+}
+
+.refresh-button-div {
+	margin-top: 8px;
+	width: 100%;
+	display: flex;
+}
+
+/* .playbookGenerationContainer {
+	font-size: 13px;
+} */
+
+/* .playbookGenerationContainer h1 {
+	padding: 5px 0 0;
+	margin: 0;
+	border: none;
+	font-weight: normal;
+	font-size: 2.7em;
+	white-space: nowrap;
+} */
+
+.playbookGenerationContainer .title {
+	border: none;
+	font-size: 2.7em;
+	font-weight: 400;
+	color: var(--vscode-foreground);
+	white-space: nowrap;
+	margin-bottom: 12px;
+}
+
+.playbookGenerationContainer .subtitle {
+	color: var(--vscode-descriptionForeground);
+	line-height: 1.4em;
+	display: block;
+	font-size: 2em;
+	margin: 0px;
+}
+
+.monaco-workbench.hc-black .part.editor>.content .playbookGenerationContainer .subtitle,
+.monaco-workbench.hc-light .part.editor>.content .playbookGenerationContainer .subtitle {
+	font-weight: 200;
+}
+
+.playbookGenerationContainer h2 {
+	font-weight: 400;
+	margin-top: 0;
+	margin-bottom: 10px;
+	font-size: 1.5em;
+	line-height: initial;
+}
+
+.index-list {
+	margin-bottom: 36px;
+}
+
+.catalogue {
+	margin-bottom: 12px;
+}
+
+.catalogue h3 {
+	margin: 0px;
+	font-weight: normal;
+}
+
+.catalogue p {
+	margin: 0px;
+}
+
+.playbookGenerationContainer a:focus {
+	outline: 1px solid -webkit-focus-ring-color;
+	outline-offset: -1px;
+	text-decoration: underline;
+}
+
+.playbookGenerationContainer .playbookGenerationSlide {
+	width: 100%;
+	height: 100%;
+	padding: 0;
+	position: absolute;
+	box-sizing: border-box;
+	left: 0;
+	top: 0;
+}
+
+.playbookGenerationContainer .playbookGenerationSlideCategories {
+	padding: 12px 24px;
+}
+
+.playbookGenerationContainer .playbookGenerationSlideCategories>.playbookGenerationCategoriesContainer>* {
+	overflow: hidden;
+	text-overflow: ellipsis;
+}
+
+.playbookGenerationContainer .playbookGenerationSlideCategories>.playbookGenerationCategoriesContainer>.categories-column>div {
+	margin-bottom: 32px;
+}
+
+.playbookGenerationContainer .playbookGenerationSlideCategories>.playbookGenerationCategoriesContainer>.categories-column-left {
+	grid-area: left-column;
+	margin-top: 25px;
+}
+
+.playbookGenerationContainer .playbookGenerationSlideCategories>.playbookGenerationCategoriesContainer>.categories-column-right {
+	grid-area: right-column;
+	margin-top: 25px;
+}
+
+.playbookGenerationContainer .playbookGenerationSlideCategories>.playbookGenerationCategoriesContainer {
+	display: grid;
+	height: 100%;
+	max-width: 1200px;
+	margin: 0 auto;
+	/* grid-template-rows: 30% minmax(min-content, auto) min-content; */
+	margin-top: 50px;
+	margin-bottom: 20px;
+	grid-template-columns: 1fr 6fr 1fr 6fr 1fr;
+	grid-template-areas:
+		".  header header header ."
+		". left-column . right-column ."
+		". footer footer footer .";
+}
+
+.playbookGenerationContainer.width-constrained .playbookGenerationSlideCategories>.playbookGenerationCategoriesContainer {
+	grid-template-rows: auto min-content minmax(min-content, auto) min-content;
+	grid-template-columns: 1fr;
+	grid-template-areas: "header" "left-column" "right-column" "footer";
+}
+
+.playbookGenerationContainer.height-constrained .playbookGenerationSlideCategories>.playbookGenerationCategoriesContainer {
+	grid-template-rows: auto minmax(min-content, auto) min-content;
+	grid-template-areas: "header" "left-column right-column" "footer footer";
+}
+
+.playbookGenerationContainer.height-constrained.width-constrained .playbookGenerationSlideCategories>.playbookGenerationCategoriesContainer {
+	grid-template-rows: min-content minmax(min-content, auto) min-content;
+	grid-template-columns: 1fr;
+	grid-template-areas: "left-column" "right-column" "footer";
+}
+
+.playbookGenerationContainer.width-constrained .playbookGenerationSlideCategories>.playbookGenerationCategoriesContainer>.header,
+.playbookGenerationContainer.height-constrained .playbookGenerationSlideCategories>.playbookGenerationCategoriesContainer>.header {
+	display: none;
+}
+
+.playbookGenerationContainer .playbookGenerationSlideCategories>.playbookGenerationCategoriesContainer>.header {
+	grid-area: header;
+	align-self: end;
+}
+
+.playbookGenerationContainer .playbookGenerationSlideCategories .gap {
+	flex: 150px 0 1000
+}
+
+.statusDisplay {
+	border-color: var(--panel-view-border);
+	border-width: thin;
+	border-style: solid;
+	border-radius: 4px;
+}
+
+.playbookGenerationContainer .playbookGenerationSlideCategories>.playbookGenerationCategoriesContainer > .header > .subtitle {
+	font-size: 1em;
+}
